Adjusted responsiveness for CategoryCard and CategoriesContainer
- CategoryCard
- CategoriesContainer
- Added global `box-sizing: border-box` rule
