+++ src/features/categories/CategoriesContainer/CategoriesContainer.module.css
@@ -1,0 +1,184 @@
+.container {
+    margin-top: 4rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 2rem 2rem 1rem;
+    gap: 2rem;
+}
+
+.container input {
+    width: 30%;
+    padding: 8px;
+    border: none;
+    border-radius: 8px;
+    font-size: 1rem;
+}
+
+.categoriesGrid {
+    display: grid;
+    width: 100%;
+    grid-template-columns: repeat(auto-fit, minmax(250px, 5fr));
+    grid-gap: 2rem;
+}
+
+@media (max-width: 500px) {
+    .container {
+        gap: 1rem;
+        padding: 1rem 3rem 1rem;
+    }
+    .container input {
+        width: 100%;
+        font-size: 14px
+    }
+}
+
+@media (min-width: 500px) and (max-width: 600px) {
+    .container {
+        margin-top: 3rem;
+        gap: 1rem;
+        padding: 1rem 1.5rem 1rem;
+    }
+    .container input {
+        width: 50%;
+        font-size: 14px
+    }
+    .categoriesGrid {
+        grid-template-columns: repeat(auto-fit, minmax(200px, 5fr));
+        grid-gap: 1rem;
+    }
+}
+
+@media (min-width: 600px) and (max-width: 800px) {
+    .container {
+        gap: 1rem;
+        padding: 1.5rem 3rem 1rem;
+    }
+    .container input {
+        width: 50%;
+        font-size: 14px
+    }
+}
+
+@media (min-width: 800px) and (max-width: 900px) {
+    .container {
+        gap: 1rem;
+        padding: 2rem 3rem 1rem;
+    }
+    .container input {
+        width: 40%;
+        font-size: 16px
+    }
+}
+
+@media (min-width: 900px) and (max-width: 1100px) {
+    .container {
+        margin-top: 3rem;
+        gap: 1rem;
+        padding: 3rem 2rem 1rem;
+    }
+    .container input {
+        width: 40%;
+        font-size: 14px;
+    }
+    .categoriesGrid {
+        grid-template-columns: repeat(auto-fit, minmax(200px, 5fr));
+        grid-gap: 1rem;
+    }
+}
+
+@media (orientation: landscape) and (min-width: 600px) and (max-width: 720px) {
+    .container {
+        margin-top: 2rem;
+        gap: 1rem;
+        padding: 1rem 1rem 1rem;
+    }
+    .container input {
+        width: 50%;
+        font-size: 12px
+    }
+    .categoriesGrid {
+        grid-template-columns: repeat(auto-fit, minmax(200px, 5fr));
+        grid-gap: 0.5rem;
+    }
+}
+
+@media (orientation: landscape) and (min-width: 720px) and (max-width: 800px) {
+    .container {
+        margin-top: 2rem;
+        gap: 1rem;
+        padding: 1rem 2rem 1rem;
+    }
+    .container input {
+        width: 40%;
+        font-size: 12px
+    }
+    .categoriesGrid {
+        grid-template-columns: repeat(auto-fit, minmax(200px, 5fr));
+        grid-gap: 1rem;
+    }
+}
+
+@media (orientation: landscape) and (min-width: 800px) and (max-width: 880px) {
+    .container {
+        margin-top: 2rem;
+        gap: 1rem;
+        padding: 1rem 3rem 1rem;
+    }
+    .container input {
+        width: 50%;
+        font-size: 12px
+    }
+    .categoriesGrid {
+        grid-template-columns: repeat(auto-fit, minmax(200px, 5fr));
+        grid-gap: 2rem;
+    }
+
+}
+@media (orientation: landscape) and (min-width: 880px) and (max-width: 1000px) {
+    .container {
+        margin-top: 2rem;
+        gap: 1rem;
+        padding: 1rem 1rem 1rem;
+    }
+    .container input {
+        width: 40%;
+        font-size: 12px
+    }
+    .categoriesGrid {
+        grid-template-columns: repeat(auto-fit, minmax(200px, 5fr));
+        grid-gap: 1rem;
+    }
+}
+
+@media (orientation: landscape) and (min-width: 1000px) and (max-width: 1150px) {
+    .container {
+        margin-top: 4rem;
+        gap: 1rem;
+        padding: 1rem 2rem 1rem;
+    }
+    .container input {
+        width: 40%;
+        font-size: 12px
+    }
+    .categoriesGrid {
+        grid-template-columns: repeat(auto-fit, minmax(220px, 5fr));
+        grid-gap: 1rem;
+    }
+}
+
+@media (orientation: landscape) and (min-width: 1150px) and (max-width: 1300px) {
+    .container {
+        margin-top: 4rem;
+        gap: 1rem;
+        padding: 1rem 2rem 1rem;
+    }
+    .container input {
+        width: 40%;
+        font-size: 14px
+    }
+    .categoriesGrid {
+        grid-template-columns: repeat(auto-fit, minmax(240px, 5fr));
+        grid-gap: 1rem;
+    }
+}
