+++ src/features/categories/CategoryCard/CategoryCard.jsx
@@ -1,4 +1,5 @@
 import { Link } from 'react-router-dom'
+import Styles from './CategoryCard.module.css'
 
 const CategoryCard = ({ category }) => {
 	const { id, name, definition, popular_drinks } = category
@@ -6,18 +7,22 @@ const CategoryCard = ({ category }) => {
 		popular_drinks[Math.floor(Math.random() * popular_drinks.length)]
 
 	return (
-		<div className="category-card">
-			<div className="image-category">
+		<div className={Styles.card}>
+			<div className={Styles.image}>
 				<img
 					src={categoryCocktail.image}
 					alt={categoryCocktail.name}
 					loading="lazy"
 				/>
-				<p>{categoryCocktail['name']}</p>
+				<div className={Styles.overlay}> {categoryCocktail.name}</div>
 			</div>
-			<h3>{name}</h3>
-			<p>{definition}</p>
-			<Link to={`/categories/${id}`} className="view-more-btn">
+			<div className={Styles.info}>
+				<Link to={`/categories/${id}`}>
+					<h3>{name}</h3>
+				</Link>
+				<p>{definition}</p>
+			</div>
+			<Link to={`/categories/${id}`} className={Styles.viewMore}>
 				View More
 			</Link>
 		</div>
