+++ src/features/cocktails/CocktailsContainer/CocktailsContainer.module.css
@@ -41,6 +41,7 @@
     flex-direction: column;
     align-items: center;
     height: 100vh;
+    margin-top: 4rem;
     padding: 0 3rem;
     gap: 2rem;
 }
